moved callbacks to an XPath-driven model
This means that Handlers (or whatever they should be called) can be
registered for specific XPath expressions in order to support
specific functionality w/o the library needing to know about it.

For example, a Handler can be registered for "/message/event" to
handle PubSub events.

A default (iq?) handler should be created to return 'not-implemented'
(or whatever the correct error message is) instead of just ignoring
stanzas.